moved gameStateSlice.ts and gameStateSlice.test.ts
